+++ examples/template-hydrogen-default/src/components/DefaultSeo.server.jsx
@@ -1,7 +1,7 @@
 import {useShopQuery} from '@shopify/hydrogen';
 import gql from 'graphql-tag';
 
-import Seo from './Seo.client';
+import DefaultSeoClient from './DefaultSeo.client';
 
 /**
  * A server component that fetches a `shop.name` and sets default values and templates for every page on a website
@@ -9,20 +9,34 @@ import Seo from './Seo.client';
 export default function DefaultSeo() {
   const {
     data: {
-      shop: {name: shopName},
+      shop: {
+        name: shopName,
+        description: shopDescription,
+        primaryDomain: {url: shopUrl},
+      },
     },
   } = useShopQuery({
     query: QUERY,
     cache: {maxAge: 60 * 60 * 12, staleWhileRevalidate: 60 * 60 * 12},
   });
 
-  return <Seo shopName={shopName} />;
+  return (
+    <DefaultSeoClient
+      shopName={shopName}
+      shopDescription={shopDescription}
+      shopUrl={shopUrl}
+    />
+  );
 }
 
 const QUERY = gql`
   query shopName {
     shop {
       name
+      description
+      primaryDomain {
+        url
+      }
     }
   }
 `;