+++ examples/template-hydrogen-default/src/pages/collections/[handle].server.jsx
@@ -10,6 +10,7 @@ import gql from 'graphql-tag';
 import LoadMoreProducts from '../../components/LoadMoreProducts.client';
 import Layout from '../../components/Layout.server';
 import ProductCard from '../../components/ProductCard';
+import CollectionSeo from '../../components/CollectionSeo.client';
 import NotFound from '../../components/NotFound.server';
 
 export default function Collection({
@@ -37,6 +38,7 @@ export default function Collection({
 
   return (
     <Layout>
+      <CollectionSeo collection={collection} />
       <h1 className="font-bold text-4xl md:text-5xl text-gray-900 mb-6 mt-6">
         {collection.title}
       </h1>
@@ -77,8 +79,13 @@ const QUERY = gql`
     collection(handle: $handle) {
       id
       title
+      description
       descriptionHtml
-
+      image {
+        url
+        width
+        height
+      }
       products(first: $numProducts) {
         edges {
           node {